Milestone 2 of ML pipeline migration

+++ core/domain/fs_services.py
@@ -19,11 +19,10 @@
 from __future__ import absolute_import  # pylint: disable=import-only-modules
 from __future__ import unicode_literals  # pylint: disable=import-only-modules
 
-import json
-
 from core.domain import fs_domain
 from core.platform import models
 import feconf
+import python_utils
 
 gae_image_services = models.Registry.import_gae_image_services()
 
@@ -91,21 +90,23 @@ def save_original_and_compressed_versions_of_image(
             micro_image_content, mimetype=mimetype)
 
 
-def save_classifier_data(exp_id, job_id, classifier_data):
+def save_classifier_data(exp_id, job_id, classifier_data_proto):
     """Store classifier model data in a file.
 
     Args:
         exp_id: str. The id of the exploration.
         job_id: str. The id of the classifier training job model.
-        classifier_data: dict. Classifier data to be stored.
+        classifier_data_proto: Object. Protobuf object of the classifier data
+            to be stored.
     """
-    filepath = '%s-classifier-data.json' % (job_id)
+    filepath = '%s-classifier-data.pb.xz' % (job_id)
     file_system_class = get_entity_file_system_class()
     fs = fs_domain.AbstractFileSystem(file_system_class(
         feconf.ENTITY_TYPE_EXPLORATION, exp_id))
+    content = python_utils.compress_to_zlib(
+        classifier_data_proto.SerializeToString())
     fs.commit(
-        filepath, json.dumps(classifier_data),
-        mimetype='application/json')
+        filepath, content, mimetype='application/octet-stream')
 
 
 def read_classifier_data(exp_id, job_id):
@@ -119,14 +120,14 @@ def read_classifier_data(exp_id, job_id):
         dict|None. The classifier data read from the file. Returns None
             if no classifier data is stored for the given job.
     """
-    filepath = '%s-classifier-data.json' % (job_id)
+    filepath = '%s-classifier-data.pb.xz' % (job_id)
     file_system_class = get_entity_file_system_class()
     fs = fs_domain.AbstractFileSystem(file_system_class(
         feconf.ENTITY_TYPE_EXPLORATION, exp_id))
     if not fs.isfile(filepath):
         return None
-    classifier_data = fs.get(filepath)
-    return json.loads(classifier_data)
+    classifier_data = python_utils.decompress_from_zlib(fs.get(filepath))
+    return classifier_data
 
 
 def delete_classifier_data(exp_id, job_id):
@@ -136,11 +137,12 @@ def delete_classifier_data(exp_id, job_id):
         exp_id: str. The id of the exploration.
         job_id: str. The id of the classifier training job model.
     """
-    filepath = '%s-classifier-data.json' % (job_id)
+    filepath = '%s-classifier-data.pb.xz' % (job_id)
     file_system_class = get_entity_file_system_class()
     fs = fs_domain.AbstractFileSystem(file_system_class(
         feconf.ENTITY_TYPE_EXPLORATION, exp_id))
-    fs.delete(filepath)
+    if fs.isfile(filepath):
+        fs.delete(filepath)
 
 
 def get_entity_file_system_class():
